perf(feed,nav,fonts,icons): server-first paint, floating bottom nav, self-hosted fonts, fixed app icons
First-paint content (server shell + initial data, like article/[id]):
- Home, Insights, Discover split into server page.tsx + client component;
  server pages prefetch via Server Actions with the default preferences and
  pass initial data; ISR (revalidate 180/300/300) serves cached HTML
- article/[id] gains revalidate=300 so article reads stop hitting MongoDB
  on every view
- PreferencesProvider and ThemeProvider no longer return null pre-mount
  (they blanked ALL server-rendered HTML); theme class is applied by a
  pre-paint inline bootstrap script, preferences hydrate from localStorage
  post-mount and trigger a refetch only when they differ from the defaults
- loading.tsx skeletons for /, /insights, /discover, /article/[id]

Fix /insights crash (root cause):
- some articles store author as a schema.org object ({"@type":"Person",
  name}), so getTrendingAuthors returned objects in id/name and React threw
  "Objects are not valid as a React child" rendering Top Journalists;
  authors are now normalised (unwrap, title-case, merge duplicates)
- Insights and Discover get explicit error states with retry instead of a
  silent empty state / infinite skeleton

Floating mobile nav:
- bottom nav is now a floating pill (inset left-4/right-4, rounded-2xl,
  blur + border + shadow) lifted above the iOS/Android home-indicator zone
  via env(safe-area-inset-bottom); viewport-fit=cover added; content
  wrapper gets matching mobile bottom padding; 48px touch targets kept

Fonts:
- render-blocking Google Fonts CSS @import replaced with next/font/google
  (self-hosted, preloaded); variables wired to the existing
  --font-sans/serif/mono tokens; dead preconnect hints removed

Icons:
- favicon.svg geometry overflowed its viewBox (clipped petals) — rebuilt
  scaled-to-fit and theme-aware via prefers-color-scheme
- app icons rebuilt with a safe zone: full-palette Seed of Life at ~70% on
  deep tanzanite (#1A0033), opaque corners (no more transparent rounded
  tile); separate purpose:any and purpose:maskable (mark inside the 80%
  safe zone) manifest entries; apple-touch-icon now opaque 180x180
