+++ src/app/article/[id]/loading.tsx
@@ -1,0 +1,5 @@
+import { ArticlePageSkeleton } from "@/components/ui/skeleton";
+
+export default function ArticleLoading() {
+  return <ArticlePageSkeleton />;
+}
