+++ src/app/discover/loading.tsx
@@ -1,0 +1,5 @@
+import { DiscoverPageSkeleton } from "@/components/ui/discover-skeleton";
+
+export default function DiscoverLoading() {
+  return <DiscoverPageSkeleton />;
+}
