+++ src/app/insights/loading.tsx
@@ -1,0 +1,5 @@
+import { InsightsPageSkeleton } from "@/components/ui/skeleton";
+
+export default function InsightsLoading() {
+  return <InsightsPageSkeleton />;
+}
