+++ src/app/layout.tsx
@@ -1,4 +1,5 @@
 import type { Metadata, Viewport } from 'next';
+import { Noto_Serif, Noto_Sans, JetBrains_Mono } from 'next/font/google';
 import './globals.css';
 import { ThemeProvider } from '@/components/theme-provider';
 import { PreferencesProvider } from '@/contexts/preferences-context';
@@ -9,8 +10,28 @@ import { OnboardingModal } from '@/components/onboarding-modal';
 import { OrganizationJsonLd, WebSiteJsonLd } from '@/components/ui/json-ld';
 import { AuthKitProvider } from '@workos-inc/authkit-nextjs/components';
 
-// Fonts are loaded via CSS @import in globals.css for better reliability
-// This avoids build failures when Google Fonts API is unreachable
+// Fonts are self-hosted via next/font (downloaded at build time, served from
+// /_next/static) — no render-blocking Google Fonts CSS request at runtime.
+// The CSS variables below are wired to the --font-sans/serif/mono theme
+// tokens in globals.css, so existing font-family rules keep working.
+const notoSerif = Noto_Serif({
+  subsets: ['latin'],
+  style: ['normal', 'italic'],
+  display: 'swap',
+  variable: '--font-noto-serif',
+});
+
+const notoSans = Noto_Sans({
+  subsets: ['latin'],
+  display: 'swap',
+  variable: '--font-noto-sans',
+});
+
+const jetbrainsMono = JetBrains_Mono({
+  subsets: ['latin'],
+  display: 'swap',
+  variable: '--font-jetbrains-mono',
+});
 
 const BASE_URL = process.env.NEXT_PUBLIC_BASE_URL || 'https://news.mukoko.com';
 
@@ -88,9 +109,9 @@ export const metadata: Metadata = {
       { url: '/favicon.svg', type: 'image/svg+xml' },
       { url: '/favicon-32.png', sizes: '32x32', type: 'image/png' },
       { url: '/favicon-16.png', sizes: '16x16', type: 'image/png' },
-      { url: '/favicon.ico' },
+      { url: '/favicon.ico', sizes: '48x48' },
     ],
-    apple: '/apple-touch-icon.png',
+    apple: [{ url: '/apple-touch-icon.png', sizes: '180x180', type: 'image/png' }],
   },
   manifest: '/manifest.json',
 };
@@ -100,6 +121,9 @@ export const viewport: Viewport = {
   initialScale: 1,
   maximumScale: 5,
   themeColor: '#4B0082',
+  // Required for env(safe-area-inset-*) to be non-zero on iOS — the floating
+  // bottom nav offsets itself above the home-indicator swipe zone with it.
+  viewportFit: 'cover',
 };
 
 export default function RootLayout({
@@ -108,11 +132,23 @@ export default function RootLayout({
   children: React.ReactNode;
 }>) {
   return (
-    <html lang="en" className="dark" suppressHydrationWarning>
+    <html
+      lang="en"
+      className={`dark ${notoSerif.variable} ${notoSans.variable} ${jetbrainsMono.variable}`}
+      suppressHydrationWarning
+    >
       <head>
-        {/* Preconnect to Google Fonts for faster font loading */}
-        <link rel="preconnect" href="https://fonts.googleapis.com" />
-        <link rel="preconnect" href="https://fonts.gstatic.com" crossOrigin="anonymous" />
+        {/* Fonts are self-hosted via next/font — no Google Fonts preconnect needed */}
+        {/* Theme bootstrap: applies the stored theme class before first paint so
+            ThemeProvider can render SSR HTML without a wrong-theme flash.
+            Static script (no interpolation); key must match ThemeProvider's
+            storageKey ("mukoko-news-theme"). */}
+        <script
+          dangerouslySetInnerHTML={{
+            __html:
+              "(function(){try{var t=localStorage.getItem('mukoko-news-theme');var d=t==='light'?false:t==='dark'?true:window.matchMedia('(prefers-color-scheme: dark)').matches;var c=document.documentElement.classList;c.remove('light','dark');c.add(d?'dark':'light');}catch(e){}})()",
+          }}
+        />
         <OrganizationJsonLd />
         <WebSiteJsonLd />
       </head>
@@ -124,7 +160,11 @@ export default function RootLayout({
             <div className="minerals-stripe" />
 
             <Header />
-            <main className="flex-1 relative z-10 pb-16 md:pb-0">{children}</main>
+            {/* Bottom padding on mobile keeps the floating nav pill from
+                covering the last row of content (nav height + lift + safe area) */}
+            <main className="flex-1 relative z-10 pb-[calc(env(safe-area-inset-bottom,0px)_+_5.5rem)] md:pb-0">
+              {children}
+            </main>
             <Footer />
             <BottomNav />
