+++ src/components/ui/skeleton.tsx
@@ -106,6 +106,46 @@ export function FeedPageSkeleton() {
   );
 }
 
+export function InsightsPageSkeleton() {
+  return (
+    <div
+      className="max-w-[1200px] mx-auto px-6 py-8"
+      aria-label="Loading insights"
+      role="status"
+      aria-live="polite"
+    >
+      {/* Header */}
+      <div className="mb-8">
+        <Skeleton className="h-9 w-48 mb-3" />
+        <Skeleton className="h-4 w-72" />
+      </div>
+
+      {/* Stats grid */}
+      <div className="grid grid-cols-3 gap-4 mb-8">
+        {Array.from({ length: 3 }).map((_, i) => (
+          <div key={i} className="bg-surface rounded-2xl p-6 border border-elevated">
+            <Skeleton className="h-8 w-8 rounded-full mx-auto mb-3" />
+            <Skeleton className="h-8 w-20 mx-auto mb-2" />
+            <Skeleton className="h-4 w-16 mx-auto" />
+          </div>
+        ))}
+      </div>
+
+      {/* Trending grid */}
+      <Skeleton className="h-6 w-40 mb-4" />
+      <div className="grid grid-cols-2 md:grid-cols-4 gap-4">
+        {Array.from({ length: 8 }).map((_, i) => (
+          <div key={i} className="bg-surface rounded-xl p-4 border border-elevated">
+            <Skeleton className="h-8 w-8 mb-3" />
+            <Skeleton className="h-5 w-24 mb-2" />
+            <Skeleton className="h-3 w-16" />
+          </div>
+        ))}
+      </div>
+    </div>
+  );
+}
+
 export function ArticlePageSkeleton() {
   return (
     <div className="max-w-3xl mx-auto px-4 py-8" aria-label="Loading article" role="status" aria-live="polite">
